# optimize the performance of binary_kernel_reduce for welford using Ro…