### Description
Optimize the performance of binary_kernel_reduce for welford using RowwiseMoments for bfloat16 on CPU.
This PR depends on the optimization of RowwiseMoments https://github.com/pytorch/pytorch/pull/84404 and https://github.com/pytorch/pytorch/pull/84405

### Testing
single socket (28cores):
```
before: torch.Size([10, 128, 10, 124]) bf16: 0.000414 s
        torch.Size([10, 128, 30, 124]) bf16: 0.00163 s
            
after: torch.Size([10, 128, 10, 124]) bf16: 2.246e-05 s
       torch.Size([10, 128, 30, 124]) bf16: 6.811e-05 s
```
single core:
```
before: torch.Size([10, 128, 10, 124]) bf16: 0.00980 s
        torch.Size([10, 128, 30, 124]) bf16: 0.0398 s
            
after: torch.Size([10, 128, 10, 124]) bf16: 0.000310 s
       torch.Size([10, 128, 30, 124]) bf16: 0.00127 s
```

cc @jgong5 @mingfeima @XiaobingSuper @sanchitintel @ashokei @jingxu10